refactor: structrue dbdir code in packages and files
