+++ dbdir/types.go
@@ -1,0 +1,28 @@
+package dbdir
+
+import (
+	"sync"
+	"time"
+
+	"github.com/jonboulle/clockwork"
+	"gorm.io/gorm"
+)
+
+type DbDirectory struct {
+	rootDir string
+	maxAge  time.Duration
+	entries map[string]*entry
+	clock   clockwork.Clock
+	models  []any
+	bigLock sync.Mutex
+}
+
+type entry struct {
+	conn            *gorm.DB
+	expirationTimer clockwork.Timer
+}
+
+type session struct {
+	gorm.Model
+	ExpiresAt time.Time
+}
