+++ dbdir_test/assert.go
@@ -1,0 +1,14 @@
+package dbdir_test
+
+import (
+	"github.com/matryer/is"
+	"gorm.io/gorm"
+)
+
+func connHasNoRows(conn *gorm.DB, is *is.I) {
+	var datas []testData
+	result := conn.Find(&datas)
+	is.NoErr(result.Error)
+
+	is.Equal(len(datas), 0)
+}
